Refactored UnitTestGenerator with PromptBuilder parameters.
