+++ cover_agent/utils.py
@@ -5,9 +5,10 @@ import re
 import yaml
 
 from typing import List
-
 from grep_ast import filename_to_lang
 
+from cover_agent.settings.config_loader import get_settings
+from cover_agent.settings.token_handling import clip_tokens, TokenEncoder
 from cover_agent.lsp_logic.utils.utils import is_forbidden_directory
 from cover_agent.version import __version__
 
@@ -150,6 +151,47 @@ def try_fix_yaml(response_text: str, keys_fix_yaml: List[str] = []) -> dict:
             pass
     except:
         pass
+
+def get_included_files(
+        included_files: list, project_root: str = "", disable_tokens=False
+    ) -> str:
+        if included_files:
+            included_files_content = []
+            file_names_rel = []
+            for file_path in included_files:
+                try:
+                    with open(file_path, "r") as file:
+                        included_files_content.append(file.read())
+                        file_path_rel = (
+                            os.path.relpath(file_path, project_root)
+                            if project_root
+                            else file_path
+                        )
+                        file_names_rel.append(file_path_rel)
+                except IOError as e:
+                    print(f"Error reading file {file_path}: {str(e)}")
+            out_str = ""
+            if included_files_content:
+                for i, content in enumerate(included_files_content):
+                    out_str += f"file_path: `{file_names_rel[i]}`\ncontent:\n```\n{content}\n```\n\n\n"
+
+            out_str = out_str.strip()
+            if not disable_tokens and get_settings().get(
+                "include_files.limit_tokens", False
+            ):
+                encoder = TokenEncoder.get_token_encoder()
+                num_input_tokens = len(encoder.encode(out_str))
+                if num_input_tokens > get_settings().get("include_files.max_tokens"):
+                    print(
+                        f"Clipping included files content from {num_input_tokens} to {get_settings().get('include_files.max_tokens')} tokens"
+                    )
+                    out_str = clip_tokens(
+                        out_str,
+                        get_settings().get("include_files.max_tokens"),
+                        num_input_tokens=num_input_tokens,
+                    )
+            return out_str
+        return ""
 
 
 def parse_args_full_repo():
