+++ tests/test_UnitTestGenerator.py
@@ -2,6 +2,7 @@ from cover_agent.CoverageProcessor import CoverageProcessor
 from cover_agent.ReportGenerator import ReportGenerator
 from cover_agent.Runner import Runner
 from cover_agent.UnitTestGenerator import UnitTestGenerator
+import cover_agent.utils
 from unittest.mock import patch, mock_open
 import datetime
 import os
@@ -19,7 +20,7 @@ class TestUnitTestGenerator:
                 mock_open(read_data="file content").return_value,
             ]
             included_files = ["invalid_file1.txt", "valid_file2.txt"]
-            result = UnitTestGenerator.get_included_files(
+            result = cover_agent.utils.get_included_files(
                 included_files, disable_tokens=True
             )
             assert (
@@ -30,7 +31,7 @@ class TestUnitTestGenerator:
     def test_get_included_files_valid_paths(self):
         with patch("builtins.open", mock_open(read_data="file content")):
             included_files = ["file1.txt", "file2.txt"]
-            result = UnitTestGenerator.get_included_files(
+            result = cover_agent.utils.get_included_files(
                 included_files, disable_tokens=True
             )
             assert (
